render_client/requirements.txt
flask==2.2.3
gunicorn==20.1.0
requests==2.28.2
python-dotenv==1.0.0
beautifulsoup4==4.12.2
python-multipart==0.0.6
pillow==9.5.0
nltk==3.8.1
tqdm==4.65.0